move modern FiatValueView
