+++ src/components/HistoryEventDetails/layouts/ExchangeFromFiatEventDetails.js
@@ -29,7 +29,7 @@ import { viewTransactionOnBlockchainAction } from 'actions/historyActions';
 import { Row, ColumnRight, Spacing } from 'components/layout/Layout';
 import Button from 'components/core/Button';
 import FeeLabel from 'components/modern/FeeLabel';
-import FiatValueView from 'components/modern/FiatValueView';
+import FiatValueView from 'components/display/FiatValueView';
 import TokenValueView from 'components/modern/TokenValueView';
 import TransactionStatusIcon from 'components/modern/TransactionStatusIcon';
 import TransactionStatusText from 'components/modern/TransactionStatusText';
