+++ src/components/display/FiatValueView.js
