+++ src/screens/Assets/components/PillarPaySummary.js
@@ -26,7 +26,7 @@ import { useTranslationWithPrefix } from 'translations/translate';
 
 // Components
 import Button from 'components/core/Button';
-import FiatValueView from 'components/modern/FiatValueView';
+import FiatValueView from 'components/display/FiatValueView';
 import Icon from 'components/core/Icon';
 import Text from 'components/core/Text';
 import Tooltip from 'components/Tooltip';
