+++ src/screens/Assets/deposits/DepositListItem.js
@@ -23,7 +23,7 @@ import { BigNumber } from 'bignumber.js';
 import styled from 'styled-components/native';
 
 // Components
-import FiatValueView from 'components/modern/FiatValueView';
+import FiatValueView from 'components/display/FiatValueView';
 import FiatChangeView from 'components/display/FiatChangeView';
 import Image from 'components/Image';
 import Text from 'components/core/Text';
