+++ src/screens/Assets/rewards/RewardListItem.js
@@ -24,7 +24,7 @@ import styled from 'styled-components/native';
 
 // Components
 import Text from 'components/core/Text';
-import FiatValueView from 'components/modern/FiatValueView';
+import FiatValueView from 'components/display/FiatValueView';
 import TokenValueView from 'components/modern/TokenValueView';
 import TokenIcon from 'components/Icons/TokenIcon';
