+++ src/screens/Assets/wallet/WalletListItem.js
@@ -24,7 +24,7 @@ import styled from 'styled-components/native';
 
 // Components
 import Text from 'components/core/Text';
-import FiatValueView from 'components/modern/FiatValueView';
+import FiatValueView from 'components/display/FiatValueView';
 import FiatChangeView from 'components/display/FiatChangeView';
 import TokenIcon from 'components/Icons/TokenIcon';
